Extracted ID3FramesParser from ID3TagParser :recycle:
