+++ Source/Parsing/Tag/ID3TagParserFactory.swift
@@ -15,17 +15,20 @@ class ID3TagParserFactory {
             tagVersionParser: ID3TagVersionParser(),
             tagPresence: ID3TagPresence(id3TagConfiguration: id3TagConfiguration),
             tagSizeParser: ID3TagSizeParser(),
-            frameSizeParser: ID3FrameSizeParser(
-                frameContentSizeParser: ID3FrameContentSizeParser(
-                    id3FrameConfiguration: id3FrameConfiguration,
-                    synchsafeIntegerDecoder: SynchsafeIntegerDecoder()),
-                id3FrameConfiguration: id3FrameConfiguration
-            ),
-            frameContentParser: ID3FrameContentParser(
-                frameContentParsingOperations: ID3FrameContentParsingOperationFactory.make(),
-                id3FrameConfiguration: id3FrameConfiguration
-            ),
-            id3TagConfiguration: id3TagConfiguration
+            id3TagConfiguration: id3TagConfiguration,
+            frameParser: ID3FramesParser(
+                frameSizeParser: ID3FrameSizeParser(
+                    frameContentSizeParser: ID3FrameContentSizeParser(
+                        id3FrameConfiguration: id3FrameConfiguration,
+                        synchsafeIntegerDecoder: SynchsafeIntegerDecoder()),
+                    id3FrameConfiguration: id3FrameConfiguration
+                ),
+                frameContentParser: ID3FrameContentParser(
+                    frameContentParsingOperations: ID3FrameContentParsingOperationFactory.make(),
+                    id3FrameConfiguration: id3FrameConfiguration
+                ),
+                id3TagConfiguration: id3TagConfiguration
+            )
         )
         return id3TagParser
     }
